% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDat.R
\name{processDat}
\alias{processDat}
\title{Converting a complex cognition dataset from horizontal format to vertical format}
\usage{
processDat(data)
}
\arguments{
\item{data}{The formatted cognition data frame containing the keys: ID, Date and var}
}
\value{
data.frame Returns a formatted data frame in vertical format that is ready for analysis
}
\description{
Converting a complex cognition dataset from horizontal format to vertical format
}
\examples{
data(dataEg1)
data_procEg1 <- processDat(dataEg1)
head(data_procEg1)

## Extended data with some new columns
data(dataEg2)
data_procEg2 <- processDat(dataEg2)
head(data_procEg2)
}
\author{
Daniel Noble - daniel.noble@unsw.edu.au
}

