% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_dat.R
\name{processDat}
\alias{processDat}
\title{Converting a complex cognition dataset from horizontal format to vertical format}
\usage{
processDat(data)
}
\arguments{
\item{data}{The formatted cognition data frame containing the keys: ID, Date and var}
}
\value{
data.frame Returns a formatted data frame in vertical format that is ready for analysis
}
\description{
Converting a complex cognition dataset from horizontal format to vertical format
}
\author{
Daniel Noble - daniel.noble@unsw.edu.au
}

